feat: xss
"Refactor Redis constants and config, enhance security measures

Updated code related to Redis configuration and constants for better naming and code clarity. RedisConstant class is removed and replaced with CacheConsts for better semantics. RedisConfig is renamed to CacheConfig to indicate that it now applies to general cache setups.

Enhanced security by introducing XssFilter and XssHttpServletRequestWrapper to prevent Cross Site Scripting(XSS) attacks. Sa-token related configurations are updated.

In addition, dependencies and pom.xml are updated, with the removal of some redundant dependencies and addition of necessary ones.

This refactoring and enhancement are part of the effort to improve code quality, security and maintainability."